requirements.txt
discord.py==1.7.3
Flask==2.0.1
googletrans==4.0.0-rc1
